Lets start making some changes, First things first Patient Registration This module or page is only available at the out patient Reception. The form will Have the following fields, 1. Photo (captures image MUST be automatically formatted to passport size.) Camera functionalities required national ID. Patient National ID number must be unique Full Name Date of Birth ( Will be used to calculate current age in other hospital forms when national ID number is called), gender Address, Phone/Cell, email, Home Hospital, District, Province, Next of kin, Kin Number, Kin Address,
Puta link to Visiting Page where visiting pateints will be registered and directedto respective department or section, Outpatient or any speciales consulation or scheduled appoitmnet. THe visiting page must also have a registaioh form link if patient if not found using national id, Once regsitared from the visiting page the patient is automatically added to the waiting list and finish the rest of the feileds on the visiting fomr.  Visiting form can also be called form the Regsition form